Update client folders
