+++ client/src/scenes/index.js
@@ -1,0 +1,9 @@
+import Play from './Play';
+import Preload from './Preload';
+import Boot from './Boot';
+
+export {
+  Play as PlayScene,
+  Preload as PreloadScene,
+  Boot as BootScene,
+}
