+++ client/src/sprites/boxes/index.js
@@ -1,0 +1,4 @@
+export { BunBox } from './BunBox'
+export { CowBox } from './CowBox'
+export { LettuceBox } from './LettuceBox'
+export { TomatoBox } from './TomatoBox'
