+++ client/src/sprites/burger/index.js
@@ -1,0 +1,7 @@
+export { BurgerBeef } from './BurgerBeef'
+export { BurgerBeefLettuce } from './BurgerBeefLettuce'
+export { BurgerBeefTomato } from './BurgerBeefTomato'
+export { BurgerBeefTomatoLettuce } from './BurgerBeefTomatoLettuce'
+export { BurgerLettuce } from './BurgerLettuce'
+export { BurgerTomato } from './BurgerTomato'
+export { BurgerTomatoLettuce } from './BurgerTomatoLettuce'
