+++ client/src/sprites/index.js
@@ -1,0 +1,25 @@
+export {
+  BunBox,
+  CowBox,
+  LettuceBox,
+  TomatoBox,
+} from './boxes'
+
+export {
+  BurgerBeef,
+  BurgerBeefLettuce,
+  BurgerBeefTomato,
+  BurgerBeefTomatoLettuce,
+  BurgerLettuce,
+  BurgerTomato,
+  BurgerTomatoLettuce,
+} from './burger'
+
+export {
+  Bun,
+  Cow,
+  Lettuce,
+  Tomato,
+} from './ingredients'
+
+export { Player } from './Player'
