+++ client/src/sprites/ingredients/index.js
@@ -1,0 +1,4 @@
+export { Bun } from './Bun'
+export { Cow } from './Cow'
+export { Lettuce } from './Lettuce'
+export { Tomato } from './Tomato'
